refactor: Extract AccountRowCaption
